% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{tv_generate_crosstab}
\alias{tv_generate_crosstab}
\title{Generate a cross-tabulation with \code{tsg}}
\usage{
tv_generate_crosstab(
  data,
  row_var,
  col_var,
  add_total = TRUE,
  add_total_row = TRUE,
  add_total_column = TRUE,
  add_percent = TRUE,
  percent_by_column = FALSE,
  convert_factor = FALSE,
  include_na = TRUE,
  position_total = c("bottom", "top"),
  output_name = "result",
  as = deparse(substitute(data))
)
}
\arguments{
\item{data}{A data frame or data.table.}

\item{row_var}{Row variable.}

\item{col_var}{Column variable.}

\item{add_total}{Add totals.}

\item{add_total_row}{Add a total row.}

\item{add_total_column}{Add a total column.}

\item{add_percent}{Add percent columns.}

\item{percent_by_column}{Calculate percentages by column.}

\item{convert_factor}{Convert labelled outputs to factors.}

\item{include_na}{Include missing values.}

\item{position_total}{Position of totals.}

\item{output_name}{Object name used in generated code for the result.}

\item{as}{Name to use in generated code.}
}
\description{
Generate a cross-tabulation with \code{tsg}
}
