% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{View an object}
\usage{
view(x, title = NULL, ...)
}
\arguments{
\item{x}{The object to display.}

\item{title}{The title to use for the display, by default
the deparsed expression is used.}

\item{...}{Passed on to the view handler.}
}
\value{
\code{x}, invisibly.
}
\description{
Queries the view handler by calling \code{\link[=get_view_handler]{get_view_handler()}}, and calls it
(if non-\code{NULL}).
}
