% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{get_view_handler}
\alias{get_view_handler}
\title{Define how to view an object}
\usage{
get_view_handler(x, ...)
}
\arguments{
\item{...}{Unused, for extensibility.}
}
\value{
A function with at least two arguments, \code{x} and \code{title}.
}
\description{
Called by \code{\link[=view]{view()}} to determine how to display an object.
The default implementation returns \code{[utils::View]}.
The RStudio IDE overrides this function, this is picked up
correctly.
}
