% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{tv_get_psgc}
\alias{tv_get_psgc}
\title{Retrieve PSGC reference data}
\usage{
tv_get_psgc(
  ...,
  token = NULL,
  version = NULL,
  level = NULL,
  harmonize = TRUE,
  minimal = TRUE,
  cols = NULL
)
}
\arguments{
\item{...}{Filter expressions forwarded to \code{phscs::get_psgc()}.}

\item{token}{Optional API token.}

\item{version}{Optional PSGC version.}

\item{level}{PSGC level.}

\item{harmonize}{Harmonize the returned data.}

\item{minimal}{Return a simplified dataset.}

\item{cols}{Optional extra columns when \code{minimal = FALSE}.}
}
\description{
Retrieve PSGC reference data
}
