% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{tv_read_rcdf}
\alias{tv_read_rcdf}
\title{Read an RCDF file}
\usage{
tv_read_rcdf(
  path,
  decryption_key,
  password = NULL,
  table = NULL,
  as = "rcdf_data"
)
}
\arguments{
\item{path}{Path to an \code{.rcdf} file.}

\item{decryption_key}{Path to the private key used to decrypt the RCDF file.}

\item{password}{Optional password for the private key.}

\item{table}{Optional table name to extract from the RCDF object.}

\item{as}{Name to use for the RCDF object in generated code.}
}
\description{
Read an RCDF file
}
